app/requirements.txt
flask==3.1.0
requests==2.32.3
python-dotenv==1.0.1
pymongo==4.11.3
beautifulsoup4==4.13.4
pymongo==4.11.3 
PyMuPDF==1.25.5
playwright==1.51.0
flask-login==0.6.3
